implement Credibility.js
