+++ src/components/Results/displayResults.js
@@ -6,32 +6,12 @@ import Typography from "@material-ui/core/Typography";
 import Textfit from "react-textfit";
 import "../../App.css";
 import ArticleSummary from './ArticleSummary';
+import Credibility from './Credibility';
 
-const useStyles = makeStyles({
-  mainGaugeHeading: {
-    fontFamily: "Graduate, sans-serif",
-    paddingTop: "20px",
-    paddingBottom: "40px",
-    textAlign: "center",
-  },
 
-  mainGauge: {},
-  resultsTable: {
-    padding: "7px",
-  },
-
-  resultsHeading: {
-    fontFamily: "Graduate, sans-serif",
-  },
-
-  resultsText: {
-    fontFamily: "Roboto, sans-serif",
-    textAlign: "center",
-  },
-});
 
 function DisplayResults(props) {
-  const classes = useStyles();
+  
   const resultsObj = JSON.parse(props.displayResults);
   const credibilityScore = resultsObj[0].score / 100;
   // accesses score from credibility analysis, assigns it to a variable,
@@ -45,70 +25,8 @@ function DisplayResults(props) {
       
       <Grid container justify="center">
       <ArticleSummary />
-        <Grid item xs={12} sm={10}>
-          <Typography variant="h5" className={classes.mainGaugeHeading}>
-            Credibility Score
-          </Typography>
-        </Grid>
-        <Grid item xs={12} sm={10} md={8} lg={6}>
-          <GaugeChart
-            className={classes.mainGauge}
-            id="main-gauge"
-            percent={credibilityScore}
-            hideText={false}
-            textColor="#000000"
-            nrOfLevels={3}
-            colors={["#ee445e", "#f8ce94", "#6bb26d"]}
-          />
-        </Grid>
-        <Grid item xs={12} sm={10} md={8} lg={8}>
-          <Grid
-            container
-            justify="center"
-            alignItems="center"
-            className={classes.resultsTable}
-          >
-            <Grid item xs={12} sm={6}>
-              <Typography
-                variant="subtitle1"
-                className={classes.resultsHeading}
-              >
-                Your URL:
-              </Typography>
-            </Grid>
-            <Grid item xs={10} sm={6}>
-              <Textfit className={classes.resultsText} mode="multi">
-                {analysedURL}
-              </Textfit>
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <Typography
-                variant="subtitle1"
-                className={classes.resultsHeading}
-              >
-                Source:
-              </Typography>
-            </Grid>
-            <Grid item xs={10} sm={6}>
-              <Typography variant="body2" className={classes.resultsText}>
-                {source}
-              </Typography>
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <Typography
-                variant="subtitle1"
-                className={classes.resultsHeading}
-              >
-                Category:
-              </Typography>
-            </Grid>
-            <Grid item xs={10} sm={6}>
-              <Typography variant="body2" className={classes.resultsText}>
-                {category}
-              </Typography>
-            </Grid>
-          </Grid>
-        </Grid>
+      <Credibility score={credibilityScore} />
+        
       </Grid>
     </div>
   );
